Crate ckb_util

Re-exports

pub use parking_lot;

Macros

shrink_to_fit

Structs

Condvar
A Condition Variable

LinkedHashMap
A linked hash map.

LinkedHashMapEntries
An insertion-order iterator over a LinkedHashMap's entries represented as an OccupiedEntry.

LinkedHashSet

Functions

lock_or_panic

Type Definitions

Mutex
A mutual exclusion primitive useful for protecting shared data

MutexGuard
An RAII implementation of a "scoped lock" of a mutex. When this structure is dropped (falls out of scope), the lock will be unlocked.

RwLock
A reader-writer lock

RwLockReadGuard
RAII structure used to release the shared read access of a lock when dropped.

RwLockWriteGuard
RAII structure used to release the exclusive write access of a lock when dropped.